+++ jgclark.Reviews/src/projects.js
@@ -6,18 +6,12 @@
 // Last updated 2026-01-24 for v1.3.0.b7, @jgclark
 //-----------------------------------------------------------------------------
 
-
 import moment from 'moment'
 import { generateProjectOutputLine } from './htmlGenerators'
 import { Project } from './projectClass'
-import {
-  finishReview,
-  renderProjectLists,
-  // updateAllProjectsListAfterChange,
-} from './reviews'
+import { finishReview, renderProjectLists } from './reviews'
 import { getReviewSettings, type ReviewConfig } from './reviewHelpers'
 import { updateAllProjectsListAfterChange } from './allProjectsListHelpers'
-// import { hyphenatedDateString } from '@helpers/dateTime'
 import { clo, JSP, logDebug, logError, logInfo, logWarn } from '@helpers/dev'
 import { archiveNoteUsingFolder } from '@helpers/NPnote'
 import { showMessageYesNo } from '@helpers/userInput'